drf_api/requirements.txt
Django>5.0,<5.2
# 파이썬은 3.13 고정
# 사용가능한버전 : 3.10, 3.11, 3.12, 3.13
djangorestframework==3.15.0,<3.15.2
django-cleanup==9.0.0
python-dotenv>=1.0.0,<2.0.0
drf-spectacular>=0.26.5,<0.27.0

psycopg2>=2.9.9,<2.10.0
PyJWT>=2.8.0,<2.9.0

pillow>=11.0.0,<11.1.0
uwsgi>=2.0.22,<2.1.0